Add intial code seperation for plugins - Allow plugins to contain products, models, stores, detail/edit/list components - Move out core dashboard mutations/getters/actions into core-store   - steve (and other stores) inherit from core-store - Add generic proxy for scenarios where `meta/proxy` doesn't work
