+++ components/FleetResources.vue
@@ -1,8 +1,8 @@
 <script>
-import { colorForState, stateDisplay, stateSort } from '@/plugins/steve/resource-instance';
 import SortableTable from '@/components/SortableTable';
 import { NAME as EXPLORER } from '@/config/product/explorer';
 import { FLEET as FLEET_ANNOTATIONS } from '@/config/labels-annotations';
+import { colorForState, stateDisplay, stateSort } from '@/plugins/core-store/resource-instance';
 
 export default {
   components: { SortableTable },
