+++ components/nav/TopLevelMenu.vue
@@ -96,7 +96,7 @@ export default {
     multiClusterApps() {
       const options = this.options;
 
-      return options.filter(opt => opt.inStore === 'management' && opt.category !== 'configuration' && opt.category !== 'legacy');
+      return options.filter(opt => (opt.inStore === 'management' || opt.isMultiClusterApp) && opt.category !== 'configuration' && opt.category !== 'legacy');
     },
 
     legacyApps() {
@@ -115,8 +115,7 @@ export default {
       const cluster = this.clusterId || this.$store.getters['defaultClusterId'];
 
       const entries = this.activeProducts.map((p) => {
-        // Try product-specific index first
-        const to = {
+        const to = p.to || {
           name:   `c-cluster-${ p.name }`,
           params: { cluster }
         };
@@ -127,13 +126,14 @@ export default {
         }
 
         return {
-          label:     this.$store.getters['i18n/withFallback'](`product."${ p.name }"`, null, ucFirst(p.name)),
-          icon:      `icon-${ p.icon || 'copy' }`,
-          value:     p.name,
-          removable: p.removable !== false,
-          inStore:   p.inStore || 'cluster',
-          weight:    p.weight || 1,
-          category:  p.category || 'none',
+          label:             this.$store.getters['i18n/withFallback'](`product."${ p.name }"`, null, ucFirst(p.name)),
+          icon:              `icon-${ p.icon || 'copy' }`,
+          value:             p.name,
+          removable:         p.removable !== false,
+          inStore:           p.inStore || 'cluster',
+          weight:            p.weight || 1,
+          category:          p.category || 'none',
+          isMultiClusterApp: p.isMultiClusterApp,
           to,
         };
       });
