+++ edit/harvesterhci.io.virtualmachinetemplateversion.vue
@@ -15,7 +15,6 @@ import { HCI } from '@/config/types';
 import { randomStr } from '@/utils/string';
 import { _CONFIG, _EDIT, _VIEW } from '@/config/query-params';
 import { HCI as HCI_ANNOTATIONS } from '@/config/labels-annotations';
-import { cleanForNew } from '@/plugins/steve/normalize';
 
 import VM_MIXIN from '@/mixins/harvester-vm';
 import CreateEditView from '@/mixins/create-edit-view';
@@ -47,7 +46,7 @@ export default {
 
   data() {
     if (this.mode === _EDIT) {
-      this.value = cleanForNew(this.value);
+      this.value = this.value.cleanForNew();
     }
 
     const templateId = this.value.templateId || this.$route.query.templateId;
@@ -168,7 +167,7 @@ export default {
         template.save();
       }
 
-      cleanForNew(this.value);
+      this.value.cleanForNew();
       this.customName = randomStr(10);
       this.$set(this.value.metadata, 'annotations', {
         ...this.value.metadata.annotations,
