+++ models/constraints.gatekeeper.sh.constraint.js
@@ -1,5 +1,4 @@
 import jsyaml from 'js-yaml';
-import { cleanForNew } from '@/plugins/steve/normalize';
 
 export const ENFORCEMENT_ACTION_VALUES = {
   DENY:   'deny',
@@ -26,7 +25,7 @@ export default {
 
   cleanForNew() {
     return () => {
-      cleanForNew(this);
+      this.$dispatch(`cleanForNew`, this);
       if (this.constraint) {
         delete this.constraint;
       }
