+++ nuxt.config.js
@@ -1,6 +1,7 @@
 import fs from 'fs';
 import path from 'path';
 
+import extensionRoutes from './plugins/app-extension/extension-routes';
 import { STANDARD } from './config/private-label';
 import { directiveSsr as t } from './plugins/i18n';
 import { trimWhitespaceSsr as trimWhitespace } from './plugins/trim-whitespace';
@@ -108,6 +109,9 @@ module.exports = {
   router: {
     base:       routerBasePath,
     middleware: ['i18n'],
+    extendRoutes(routes, resolve) {
+      routes.push(...extensionRoutes.routes(resolve));
+    }
   },
 
   build: {
@@ -283,12 +287,14 @@ module.exports = {
     '@nuxtjs/webpack-profile',
     'cookie-universal-nuxt',
     'portal-vue/nuxt',
-    '~/plugins/steve/rehydrate-all',
+    '~/plugins/core-store/rehydrate-all',
     '@nuxt/content',
   ],
 
   // Vue plugins
   plugins: [
+    // '~/plugins/app-extension/epinio',
+
     // Third-party
     '~/plugins/axios',
     '~/plugins/tooltip',
@@ -320,6 +326,7 @@ module.exports = {
     '/apis':         proxyWsOpts(api), // Management k8s API
     '/v1':           proxyWsOpts(api), // Management Steve API
     '/v3':           proxyWsOpts(api), // Rancher API
+    '/proxy':        genericProxy(api), // api can be anything
     '/v3-public':    proxyOpts(api), // Rancher Unauthed API
     '/api-ui':       proxyOpts(api), // Browser API UI
     '/meta':         proxyOpts(api), // Browser API UI
@@ -355,6 +362,19 @@ module.exports = {
 
   typescript: { typeCheck: { eslint: { files: './**/*.{ts,js,vue}' } } }
 };
+
+function genericProxy(target) {
+  return {
+    target,
+    secure:       !dev,
+    pathRewrite:  { '^/proxy': '' },
+    changeOrigin: true,
+    router(req) {
+      return req.headers['x-api-host'];
+    },
+    onError
+  };
+}
 
 function proxyOpts(target) {
   return {
