+++ plugins/core-store/resource-instance.js
@@ -35,8 +35,8 @@ import {
   AS, _YAML, MODE, _CLONE, _EDIT, _VIEW, _UNFLAG, _CONFIG
 } from '@/config/query-params';
 
-import { SELF } from '@/plugins/steve/resource-proxy';
-import { cleanForNew, normalizeType } from './normalize';
+import { SELF } from '@/plugins/core-store/resource-proxy';
+import { normalizeType } from './normalize';
 
 const STRING_LIKE_TYPES = [
   'string',
@@ -795,7 +795,7 @@ export default {
         throw new Error(`Unknown link ${ linkName } on ${ this.type } ${ this.id }`);
       }
 
-      return this.$dispatch('request', opt);
+      return this.$dispatch('request', { opt, type: this.type });
     };
   },
 
@@ -837,7 +837,7 @@ export default {
       opt.headers['content-type'] = 'application/json-patch+json';
       opt.data = data;
 
-      return this.$dispatch('request', opt);
+      return this.$dispatch('request', { opt, type: this.type });
     };
   },
 
@@ -912,7 +912,7 @@ export default {
       }
 
       try {
-        const res = await this.$dispatch('request', opt);
+        const res = await this.$dispatch('request', { opt, type: this.type });
 
         // console.log('### Resource Save', this.type, this.id);
 
@@ -945,7 +945,7 @@ export default {
 
       opt.method = 'delete';
 
-      const res = await this.$dispatch('request', opt);
+      const res = await this.$dispatch('request', opt, { opt, type: this.type });
 
       if ( res?._status === 204 ) {
         // If there's no body, assume the resource was immediately deleted
@@ -1183,7 +1183,7 @@ export default {
         const obj = jsyaml.load(yaml);
 
         if (mode !== 'edit') {
-          cleanForNew(obj);
+          this.$dispatch(`cleanForNew`, obj);
         }
 
         if (obj._type) {
@@ -1201,7 +1201,7 @@ export default {
 
   cleanForNew() {
     return () => {
-      cleanForNew(this);
+      this.$dispatch(`cleanForNew`, this);
     };
   },
