+++ store/type-map.js
@@ -94,6 +94,7 @@
 //                               resource: undefined       -- Use this resource in ResourceDetails instead
 //                               resourceDetail: undefined -- Use this resource specifically for ResourceDetail's detail component
 //                               resourceEdit: undefined   -- Use this resource specifically for ResourceDetail's edit component
+//                               customRoute: undefined,
 //                           }
 // )
 // ignoreGroup(group):        Never show group or any types in it
@@ -127,7 +128,7 @@ import {
 
 import { NAME as EXPLORER } from '@/config/product/explorer';
 import isObject from 'lodash/isObject';
-import { normalizeType } from '@/plugins/steve/normalize';
+import { normalizeType } from '@/plugins/core-store/normalize';
 import { sortBy } from '@/utils/sort';
 import { haveV1Monitoring, haveV2Monitoring } from '@/utils/monitoring';
 
@@ -211,8 +212,10 @@ export function DSL(store, product, module = 'type-map') {
       store.commit(`${ module }/configureType`, { ...options, match });
     },
 
-    componentForType(match, replace) {
-      store.commit(`${ module }/componentForType`, { match, replace });
+    componentForType(match, replace, plugin) {
+      store.commit(`${ module }/componentForType`, {
+        match, replace, plugin
+      });
     },
 
     ignoreType(regexOrString) {
@@ -291,11 +294,10 @@ export async function applyProducts(store) {
   }
 
   called = true;
-  const ctx = require.context('@/config/product', true, /.*/);
+  const builtIn = require.context('@/config/product', true, /.*/);
+  const builtInProducts = builtIn.keys().filter(path => !path.endsWith('.js')).map(path => path.substr(2));
 
-  const products = ctx.keys().filter(path => !path.endsWith('.js')).map(path => path.substr(2));
-
-  for ( const product of products ) {
+  for ( const product of builtInProducts ) {
     const impl = await loadProduct(product);
 
     if ( impl?.init ) {
@@ -423,6 +425,7 @@ export const getters = {
       showAge:     true,
       canYaml:     true,
       namespaced:  null,
+      customRoute: undefined,
     };
 
     return (schemaOrType) => {
@@ -768,6 +771,7 @@ export const getters = {
           count:       count ? count.summary.count || 0 : null,
           byNamespace: count ? count.namespaces : {},
           revision:    count ? count.revision : null,
+          route:       typeOptions.customRoute
         };
       }
 
@@ -984,7 +988,12 @@ export const getters = {
       }
 
       try {
-        require.resolve(`@/detail/${ key }`);
+        if ( key.plugin) {
+          require.resolve(`@/plugins/app-extension/${ key.plugin }/detail/${ key.type }`);
+        } else {
+          require.resolve(`@/detail/${ key }`);
+        }
+
         cache[key] = true;
       } catch (e) {
         cache[key] = false;
@@ -1005,7 +1014,12 @@ export const getters = {
       }
 
       try {
-        require.resolve(`@/edit/${ key }`);
+        if ( key.plugin) {
+          require.resolve(`@/plugins/app-extension/${ key.plugin }/edit/${ key.type }`);
+        } else {
+          require.resolve(`@/edit/${ key }`);
+        }
+
         cache[key] = true;
       } catch (e) {
         cache[key] = false;
@@ -1107,7 +1121,11 @@ export const getters = {
       });
 
       if ( mapping ) {
-        out = mapping.replace;
+        if (mapping.plugin) {
+          out = { plugin: mapping.plugin, type };
+        } else {
+          out = mapping.replace;
+        }
       } else if ( subType ) {
         // Try again without the subType
         out = getters.componentFor(type);
@@ -1419,10 +1437,12 @@ export const mutations = {
     _addMapping(state.typeMoveMappings, match, group, weight);
   },
 
-  componentForType(state, { match, replace }) {
+  componentForType(state, { match, replace, plugin }) {
     match = ensureRegex(match);
     match = regexToString(match);
-    state.typeToComponentMappings.push({ match, replace });
+    state.typeToComponentMappings.push({
+      match, replace, plugin
+    });
   },
 
   configureType(state, options) {
