+++ tsconfig.json
@@ -15,12 +15,12 @@
     "noEmit": true,
     "baseUrl": ".",
     "paths": {
+      "@/*": [
+        "./*"
+      ],
       "~/*": [
         "./*"
       ],
-      "@/*": [
-        "./*"
-      ]
     },
     "types": [
       "@types/node",
