+++ utils/extensions.js
@@ -1,0 +1,1 @@
+export const EXTENSION_PREFIX = 'ext';
